Format code a bit
